turnier/backend/requirements.txt
flask-restx
flask
flask-cors
urllib3
schedule
flask_sqlalchemy
psycopg2
waitress
requests
werkzeug
